Get the delete data page working.
