+++ src/settings/_index.njk
@@ -21,21 +21,6 @@
   </h2>
 
   <h3>
-    Delete Data
-  </h3>
-  <p>
-    If you want to delete all your data, you can do that. Just be warned, this is
-    irreversible, so you may want to export your data first, just in case you
-    change your mind later.
-  </p>
-  <div>
-    <button type="button" data-delete-data>
-      Delete All Data
-    </button>
-  </div>
-  <div data-delete-progress aria-live="polite"></div>
-
-  <h3>
     Bulk Upload from CSV
   </h3>
   <p>
