+++ src/settings/delete-data.njk
@@ -15,8 +15,21 @@
       Go Back
     </a>
   </div>
+
+  <p>
+    If you want to delete all your data, you can do that. Just be warned, this
+    is irreversible, so you may want to
+    <a href="/settings/export-data">export your data</a>
+    first, just in case you change your mind later.
+  </p>
+  <div>
+    <button type="button" data-delete-data>
+      Delete All Data
+    </button>
+  </div>
+  <div data-delete-progress aria-live="polite"></div>
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/settings/delete-data.js"></script>
+  <script src="/scripts/settings/delete-data.mjs" type="module"></script>
 {% endblock %}
